+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.2.0
+- Use ember-auto-import instead of ember-cli-node-assets
+- Use peerDependencies and validate them
+
+## 2.1.0
+
+- Bump `ember-cli-babel` to v7
+
 ## 2.0.0
 
 - Bump _all_ dependencies to latest